third_party/WebKit/LayoutTests/inspector/layers/layer-canvas-log-expected.txt

Canvas log: 
{
    0 : {
        name : "drawRect"
    }
    1 : {
        name : "drawRect"
    }
}
